Add function to rescale parameters.
I decide to still use this repository although I rewrote all the code. I think this is because I want to use the powers of git and its documentation history.
I modified all the program (so far) to make it simple to read parameters from a file -- as simpler as C++ allows. I am now actually using a **struct** instead of a class to hold the parameters. This is more intuitive.
There is now a function that beautifully prints the parameters to the terminal. It includes colors! The print function I am sure that is working fine, as well as the func that read input params.
The functions to rescale parameters and calculate derived parameters need to be tested, if this is possible at all.