Add function to set initial positions.
Finally I made some progress in C++! I think the initial positions are working, but it would be nice to visualize it, however I need to learn some good, modern visualization tool.
After finishing the initialization functions, perhaps I should comment and document it.
I also need to write a unique function to call the initializations functions of the simulations, then it will be simples in the main.cpp.